retail-store-inventory/retail-usecase/visualization/telegraf/src/requirements.txt
shapely==2.0.1
requests==2.28.2
influxdb==5.3.1
eclipse-zenoh==0.10.0rc0
pyyaml==6.0.1
fire==0.5.0
